# contracts/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_submit_feedback_api_v1_feedback_post": {
        "properties": {
          "images": {
            "default": [],
            "items": {
              "contentMediaType": "application/octet-stream",
              "type": "string"
            },
            "title": "Images",
            "type": "array"
          },
          "message": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "module": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Module",
            "type": "string"
          },
          "page_path": {
            "anyOf": [
              {
                "maxLength": 512,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Path"
          },
          "suggestion": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggestion"
          }
        },
        "required": [
          "module",
          "message"
        ],
        "title": "Body_submit_feedback_api_v1_feedback_post",
        "type": "object"
      },
...
      "FeedbackAttachmentView": {
        "properties": {
          "content_type": {
            "title": "Content Type",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "content_type",
          "size_bytes"
        ],
        "title": "FeedbackAttachmentView",
        "type": "object"
      },
      "FeedbackView": {
        "properties": {
          "attachments": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/FeedbackAttachmentView"
            },
            "title": "Attachments",
            "type": "array"
          },
          "author_name": {
            "title": "Author Name",
            "type": "string"
          },
          "category": {
            "title": "Category",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "module": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Module"
          },
          "page_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Path"
          },
          "suggestion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggestion"
          }
        },
        "required": [
          "id",
          "author_name",
          "category",
          "message",
          "created_at"
        ],
        "title": "FeedbackView",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Digital Worker Platform API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/feedback": {
      "get": {
        "operationId": "list_feedback_api_v1_feedback_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FeedbackView"
                  },
                  "title": "Response List Feedback Api V1 Feedback Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Feedback",
        "tags": [
          "feedback"
        ]
      },
      "post": {
        "operationId": "submit_feedback_api_v1_feedback_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_submit_feedback_api_v1_feedback_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Feedback",
        "tags": [
          "feedback"
        ]
      }
    },
    "/api/v1/feedback/{feedback_id}/attachments/{attachment_id}": {
      "get": {
        "description": "The screenshot bytes, for the inbox. Same scope as the inbox itself.",
        "operationId": "read_attachment_api_v1_feedback__feedback_id__attachments__attachment_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "feedback_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Feedback Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "attachment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Attachment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Attachment",
        "tags": [
          "feedback"
        ]
      }
    },
...
  }
}